refactor(experimental): set `__DEV__` earlier in the test environment ## Summary
This code used to define `__DEV__` before the first test ran, but if `__DEV__` was needed _before_ that (eg. in a module factory of something under test) it would not be defined yet. This PR defines it right off the bat, in addition to resetting it before each test.

## Test Plan

Add `__DEV__` into a module factory of something under test. No more runtime fatal.