refactor(experimental): a strategy for expiring a transaction confirmation when the block height is exceeded (#1699)
# Summary

Transactions with a lifetime as long as their blockhash is valid can be thought to expire after that blockhash becomes invalid. This strategy watches the progression of slots and throws an error after exceeding the height for which the blockhash is valid.

# Test Plan

```
cd packages/library
pnpm test:unit:browser
pnpm test:unit:node
```