refactor(experimental): create a sham for `PublicKey` (#1861)
# Preamble

Imagine that you depend on a module that itself depends on the legacy `@solana/web3.js`. This makes it difficult for you to remove the legacy library from your project.

The chances are, though, that module only needs a _fraction_ of what web3.js provides.

The goal of the sham – of which this PR is the first – is to provide drop-in replacements for just the things that a set list of known dependencies require. You should be able to hot-swap the sham in via npm `overrides` or yarn `resolutions` and thereby drop the legacy web3.js from your bundle.

# Summary

This PR introduces an interface compatible(-ish) `PublicKey` that you can use as a drop-in replacement for the one in `@solana/web3.js@1`

# Test Plan

```
cd packages/library-legacy-sham/
pnpm test:unit:browser
pnpm test:unit:node
```

See later diffs for a functional test plan where we actually drop the sham in to an actual application and see the results.

Addresses #1825.