refactor(experimental): remove fallbackSender when signing and sending transactions (#1874)
This PR introduces an `ITransactionWithSingleSendingSigner` opaque type that extends `ITransactionWithSigners` whose responsibility is to ensure there is a single `TransactionSendingSigner` to choose from inside the transaction.

It offers a couple of type guards `isTransactionWithSingleSendingSigner` and `assertIsTransactionWithSingleSendingSigner` that fail if:
- There is no `TransactionSendingSigner` on the transaction or
- There are multiple `TransactionSendingSigner` such that at least more than one of them do not implement other transaction signing interfaces.

Additionally it removes the `fallbackSender` option in favour of requiring a `ITransactionWithSingleSendingSigner` type. This means, you now have to explicitly assert the transaction has the appropriate state before feeding it to the `signAndSendTransactionWithSigner` helper function.

```ts
assertIsTransactionWithSingleSendingSigner(myTransaction);
const myTransactionSignature = await signAndSendTransactionWithSigners(myTransaction);
```

Alternatively, you may use the `isTransactionWithSingleSendingSigner()` function to provide a fallback in case the transaction does not contain any sending signer.

```ts
let transactionSignature: SignatureBytes;
if (isTransactionWithSingleSendingSigner(transaction)) {
    transactionSignature = await signAndSendTransactionWithSigners(transaction);
} else {
    const signedTransaction = await signTransactionWithSigners(transaction);
    const encodedTransaction = getBase64EncodedWireTransaction(signedTransaction);
    transactionSignature = await rpc.sendTransaction(encodedTransaction).send();
}
```

This improvement was suggested by the reviews of https://github.com/solana-labs/solana-web3.js/pull/1792.