refactor(experimental): add blockhash codec and type assertions
Similar to the `AddressCodec` in `@solana/addresses`, we could use a typed
32-byte codec that can chain base64 encoding into its serialization, for
`Blockhash`.

Here I've added the same kind of combined codec as `getAddressCodec` to the
`blockhash` type in `@solana/rpc-types`.

Note: The implementation is almost identical, with the exception of the types,
and the tests are exactly the same as `getAddressCodec`.

I've also added some more assertions to `blockhash` to keep the API consistent.