Add an encoder for the new signatures map (#2462)
This will be used for the new transaction encoder

We encode only the signature bytes in order, filling missing (null) signature bytes with an all-0 signature. 

Note that signatures cannot be decoded in isolation, we need the account addresses from the transaction message to recreate the signatures map. So this will only be an encoder, and decoding will only be available at the transaction level (which has access to the message).