Move decompile-transaction to decompile-message in transaction-messages (#2465)
This PR adds decompile-message to the transaction-messages package. This is the inverse of compile-message: it takes a compiled message (the form we serialise) and returns a TransactionMessage. I think this completes the transaction messages package, other than renaming stuff! 

The code and tests come from decompile-transaction, which is moved. Changes are very minimal, just renaming things and removing everything to do with signatures

This was used by the previous transaction deserializer. I've inlined the signatures code into this, and refactored it to use decompile-message and then handle signatures itself. This is all just temporary since this previous transaction serializer will be removed.

I've also modified the library `decodeTransaction`, which uses an RPC to fetch address lookup tables and then uses them to decompile the transaction message, to use the previous `getTransactionDecoder`. This will keep its behaviour unchanged for now. I think this will be refactored to `decodeTransactionMessage` that takes message bytes + an RPC and returns a transaction message object. But that's for another PR!