A `useSignIn()` hook for accessing the Sign In With Solana feature (#2928)
# Summary

This hook returns essentially the raw sign in function from the `solana:signIn` feature. It upcasts the returned `WalletAccount` to a `UiWalletAccount` usable with the rest of the `@solana/react` API.

Callers are responsible for verifying the signed message and its signature.

# Test plan

```
pnpm turbo test:unit:node test:unit:browser test:typecheck
```